Criando contexto por toda aplicação
